airside_ops_system/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.36
WTForms==3.1.1
Werkzeug==3.1.3
python-dotenv==1.1.1
bcrypt==4.2.1
PyJWT==2.8.0
celery==5.4.0
redis==5.2.0
reportlab==4.2.5
pandas==2.2.3
openpyxl==3.1.5
Pillow==11.0.0
python-dateutil==2.8.2
gevent==24.11.1
gunicorn==21.2.0
psycopg2-binary==2.9.10
eventlet==0.37.0
pyotp==2.9.0
qrcode==7.4.2
bleach==6.1.0
email-validator==2.1.0
